+++ builds/8chanSS.meta.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         8chanSS
-// @version      1.50.3
+// @version      1.51.0
 // @namespace    8chanss
 // @description  A userscript to add functionality to 8chan.
 // @author       otakudude
@@ -15,6 +15,9 @@
 // @grant        GM.setValue
 // @grant        GM.deleteValue
 // @grant        GM.listValues
+// @grant        GM.xmlHttpRequest
+// @connect      youtube.com
+// @connect      i.ytimg.com
 // @run-at       document-start
 // @updateURL    https://github.com/otacoo/8chanSS/releases/latest/download/8chanSS.meta.js
 // @downloadURL  https://github.com/otacoo/8chanSS/releases/latest/download/8chanSS.user.js